ContextIQ Pro/requirements.txt
# API & server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# AI
openai>=1.30.0

# Vector DB
qdrant-client>=1.9.0

# Web scraping
playwright>=1.44.0
beautifulsoup4>=4.12.0

# Document parsing
pymupdf>=1.24.0          # PDF (fitz)
python-docx>=1.1.0       # DOCX

# Text splitting
langchain-text-splitters>=0.2.0

# Web search
requests>=2.31.0

# Config
python-dotenv>=1.0.0

# UI
gradio>=4.36.0